requirements.txt
annotated-types==0.7.0
asgiref==3.9.1
black==25.1.0
cffi==1.17.1
click==8.2.1
colorama==0.4.6
cryptography==45.0.7
Django==5.2.6
django-cors-headers==4.7.0
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
flake8==7.3.0
iniconfig==2.1.0
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.4.0
pluggy==1.6.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytokens==0.3.0
redis==6.4.0
six==1.17.0
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
